fix: handle GitHub App update redirects and deduplicate sync logic
GitHub's setup URL fires for both installs and updates (e.g. repository
selection changes), but the callback only handled the install case. When
state was missing on update redirects, the flow errored out.

- Extract shared `syncGithubInstallation` helper to deduplicate sync
  logic between initial-sync and manual sync routes, and add stale repo
  cleanup
- Handle stateless update redirects by falling back to session + existing
  installation lookup
- Add "Manage Installation" button so users can update repo selection
- Return `github_updated` success message for update flows
- Support direct fetch for dev mode (bypasses QStash)
- Add RFC for longer-term webhook-first architecture
