+++ apps/api/src/app/api/github/sync-installation.ts
@@ -1,0 +1,219 @@
+import { db } from "@superset/db/client";
+import {
+	githubInstallations,
+	githubPullRequests,
+	githubRepositories,
+} from "@superset/db/schema";
+import { subDays } from "date-fns";
+import { and, eq, notInArray } from "drizzle-orm";
+
+import { githubApp } from "./octokit";
+
+interface SyncGithubInstallationParams {
+	installationDbId: string;
+	githubInstallationId: string;
+	organizationId: string;
+	logPrefix: string;
+}
+
+export async function syncGithubInstallation({
+	installationDbId,
+	githubInstallationId,
+	organizationId,
+	logPrefix,
+}: SyncGithubInstallationParams) {
+	const octokit = await githubApp.getInstallationOctokit(
+		Number(githubInstallationId),
+	);
+
+	const repos = await octokit.paginate(
+		octokit.rest.apps.listReposAccessibleToInstallation,
+		{ per_page: 100 },
+	);
+
+	console.log(`[${logPrefix}] Found ${repos.length} repositories`);
+
+	const accessibleRepoIds = repos.map((repo) => String(repo.id));
+
+	if (accessibleRepoIds.length === 0) {
+		await db
+			.delete(githubRepositories)
+			.where(eq(githubRepositories.installationId, installationDbId));
+	} else {
+		await db
+			.delete(githubRepositories)
+			.where(
+				and(
+					eq(githubRepositories.installationId, installationDbId),
+					notInArray(githubRepositories.repoId, accessibleRepoIds),
+				),
+			);
+	}
+
+	for (const repo of repos) {
+		await db
+			.insert(githubRepositories)
+			.values({
+				installationId: installationDbId,
+				organizationId,
+				repoId: String(repo.id),
+				owner: repo.owner.login,
+				name: repo.name,
+				fullName: repo.full_name,
+				defaultBranch: repo.default_branch ?? "main",
+				isPrivate: repo.private,
+			})
+			.onConflictDoUpdate({
+				target: [githubRepositories.repoId],
+				set: {
+					installationId: installationDbId,
+					organizationId,
+					owner: repo.owner.login,
+					name: repo.name,
+					fullName: repo.full_name,
+					defaultBranch: repo.default_branch ?? "main",
+					isPrivate: repo.private,
+					updatedAt: new Date(),
+				},
+			});
+	}
+
+	const thirtyDaysAgo = subDays(new Date(), 30);
+
+	for (const repo of repos) {
+		const [dbRepo] = await db
+			.select()
+			.from(githubRepositories)
+			.where(eq(githubRepositories.repoId, String(repo.id)))
+			.limit(1);
+
+		if (!dbRepo) continue;
+
+		const prs: Awaited<ReturnType<typeof octokit.rest.pulls.list>>["data"] = [];
+
+		for await (const response of octokit.paginate.iterator(
+			octokit.rest.pulls.list,
+			{
+				owner: repo.owner.login,
+				repo: repo.name,
+				state: "all",
+				sort: "updated",
+				direction: "desc",
+				per_page: 100,
+			},
+		)) {
+			let reachedCutoff = false;
+			for (const pr of response.data) {
+				if (new Date(pr.updated_at) < thirtyDaysAgo) {
+					reachedCutoff = true;
+					break;
+				}
+				prs.push(pr);
+			}
+			if (reachedCutoff) break;
+		}
+
+		console.log(
+			`[${logPrefix}] Found ${prs.length} PRs (last 30 days) for ${repo.full_name}`,
+		);
+
+		for (const pr of prs) {
+			const { data: checksData } = await octokit.rest.checks.listForRef({
+				owner: repo.owner.login,
+				repo: repo.name,
+				ref: pr.head.sha,
+			});
+
+			const checks = checksData.check_runs.map(
+				(c: (typeof checksData.check_runs)[number]) => ({
+					name: c.name,
+					status: c.status,
+					conclusion: c.conclusion,
+					detailsUrl: c.details_url ?? undefined,
+				}),
+			);
+
+			let checksStatus = "none";
+			if (checks.length > 0) {
+				const hasFailure = checks.some(
+					(c: {
+						name: string;
+						status: string;
+						conclusion: string | null;
+						detailsUrl?: string;
+					}) => c.conclusion === "failure" || c.conclusion === "timed_out",
+				);
+				const hasPending = checks.some(
+					(c: {
+						name: string;
+						status: string;
+						conclusion: string | null;
+						detailsUrl?: string;
+					}) => c.status !== "completed",
+				);
+
+				checksStatus = hasFailure
+					? "failure"
+					: hasPending
+						? "pending"
+						: "success";
+			}
+
+			await db
+				.insert(githubPullRequests)
+				.values({
+					repositoryId: dbRepo.id,
+					organizationId,
+					prNumber: pr.number,
+					nodeId: pr.node_id,
+					headBranch: pr.head.ref,
+					headSha: pr.head.sha,
+					baseBranch: pr.base.ref,
+					title: pr.title,
+					url: pr.html_url,
+					authorLogin: pr.user?.login ?? "unknown",
+					authorAvatarUrl: pr.user?.avatar_url ?? null,
+					state: pr.state,
+					isDraft: pr.draft ?? false,
+					additions: 0,
+					deletions: 0,
+					changedFiles: 0,
+					reviewDecision: null,
+					checksStatus,
+					checks,
+					mergedAt: pr.merged_at ? new Date(pr.merged_at) : null,
+					closedAt: pr.closed_at ? new Date(pr.closed_at) : null,
+				})
+				.onConflictDoUpdate({
+					target: [
+						githubPullRequests.repositoryId,
+						githubPullRequests.prNumber,
+					],
+					set: {
+						organizationId: dbRepo.organizationId,
+						headSha: pr.head.sha,
+						title: pr.title,
+						state: pr.state,
+						isDraft: pr.draft ?? false,
+						checksStatus,
+						checks,
+						mergedAt: pr.merged_at ? new Date(pr.merged_at) : null,
+						closedAt: pr.closed_at ? new Date(pr.closed_at) : null,
+						lastSyncedAt: new Date(),
+						updatedAt: new Date(),
+					},
+				});
+		}
+	}
+
+	await db
+		.update(githubInstallations)
+		.set({ lastSyncedAt: new Date() })
+		.where(eq(githubInstallations.id, installationDbId));
+
+	console.log(`[${logPrefix}] Sync completed successfully`);
+
+	return {
+		repositoriesCount: repos.length,
+	};
+}
